refactor: header and footer change to components
