+++ src/pages/auth/ChangePassword.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/components/Form'
 import { Input } from '@/components/Input'
 import { Button } from '@components/Button'
-import { Typography } from '@components/Typograpy.tsx'
+import { Typography } from '@components/Typograpy'
 import { Eye, EyeOff } from 'lucide-react'
 
 const formSchema = z
